# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mycorrhizal CRM API
  description: >
    REST API for Mycorrhizal's contact model. As of WP-71 (P2,
    docs/fork-plan/50-integration-and-rebrand.md), the contact endpoints speak
    the neutral RFC 9553-shaped model (contactmodel.Record/Card) rather than
    the legacy flat Contact DTO. GET /contacts (list) returns a slim
    ContactSummary per item; GET/POST/PUT /contacts/{id} exchange the full
    Card/CRM shape (ContactRecordResponse / ContactRecordInput) — a
    symmetric read/write contract.


    T8 (docs/fork-plan/tickets/16-T8-openapi.md) extended this spec to cover
    the whole REST surface: every route registered by
    backend/routes/routes.go under /api/v1 plus the unversioned /health. The
    drift test (backend/openapi_test.go) enumerates the live Gin router and
    fails on any registered route with no spec entry, so this file cannot go
    stale again.

    Coverage notes:

    - **Auth:** every path under /api/v1 except the public auth group
      (register, login, logout, check-password-strength, password-reset,
      auth/oidc/config) requires a session cookie (`auth_token`, set by login
      and OIDC callback) or an `Authorization: Bearer <token>` header. API
      tokens (mycorrhizal_…) are accepted on the general REST path only when
      their scope is `full`; a `carddav`-scoped token is rejected with 403.
      API tokens cannot access the admin paths at all.
...
      `/auth/oidc/config` is always present.
...
      them.
...
servers:
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /tags:
    get:
      summary: List tags
      operationId: listTags
      description: >
        Cursor-paginated (T17). Tags are a hard-delete collection — the
        sync.mode is full_resync: a client re-pulls the collection rather than
        using ?since= (tags have no tombstone).
      parameters:
        - name: cursor
          in: query
          description: Opaque resume token from the previous page's next_cursor.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, default: 25 }
        - name: order
          in: query
          description: Direction for the (updated_at, id) cursor order.
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: include_contacts
          in: query
          description: When true, also returns a contacts array (all contact_tags join rows for the page's tags).
          schema: { type: boolean, default: false }
      responses:
        "401":
          description: Not authenticated (missing or invalid session cookie or full-scope bearer token).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "200":
          description: A page of tags.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tag"
                  total: { type: integer }
                  next_cursor:
                    type: string
                    description: Resume token for the next page; empty when there are no more rows.
                  limit: { type: integer }
                  sync:
                    $ref: "#/components/schemas/SyncInfo"
                  contacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContactTag"
    post:
      summary: Create a tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "200":
          description: Tag created (200, not 201).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Tag created successfully" }
                  tag:
                    $ref: "#/components/schemas/Tag"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tags/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a tag with its contacts
      operationId: getTag
      responses:
        "200":
          description: The tag and its tagging join rows.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tag:
                    $ref: "#/components/schemas/Tag"
                  contacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContactTag"
        "404":
          description: Tag not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a tag
      operationId: updateTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "200":
          description: The updated tag (raw, not wrapped).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Tag not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a tag (tagging cascade)
      operationId: deleteTag
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Tag not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tags/{id}/contacts:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Tag a contact
      operationId: addContactTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactTagInput"
      responses:
        "200":
          description: Contact tagged.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Contact tagged" }
                  tagging:
                    $ref: "#/components/schemas/ContactTag"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Tag or contact not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The contact is already tagged with this tag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tags/{id}/contacts/{vcard_uid}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: vcard_uid
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: Remove a tag from a contact
      operationId: removeContactTag
      responses:
        "200":
          description: Tagging removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Tag or tagging not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # --- Custom field definitions ---
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: auth_token
      description: The httpOnly session cookie set by POST /login and the OIDC callback.
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A JWT or a `mycorrhizal_…` API token. On the general REST path a
        carddav-scoped token is rejected with 403 (only `full`-scope tokens
        work); API tokens are rejected on every /admin path entirely.
  schemas:
...
    Error:
      type: object
      description: The standard error envelope produced by apperrors.AbortWithError.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
        request_id: { type: string }
        timestamp: { type: string, format: date-time }
    SyncInfo:
      type: object
      description: >
        T17 sync-mode contract carried by every cursor-paginated list
        response. `mode` is how THIS collection syncs; the two arrays are the
        static map of every syncable collection, so a client can plan a full
        sync without hardcoding endpoint names. Incremental collections are
        soft-deleted, support ?since= change feeds, and drop `total`;
        full_resync collections are hard-deleted, are re-pulled wholesale, and
        keep `total`.
      properties:
        mode: { type: string, enum: [incremental, full_resync] }
        incremental:
          type: array
          items: { type: string }
          description: "Collections that sync incrementally via ?since=: contacts, notes, activities, life_events, preferences."
        full_resync:
          type: array
          items: { type: string }
          description: "Collections that require a full re-pull: circles, households, tags, relationship_edges, field_definitions."
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
      additionalProperties: true
...
    Tag:
      type: object
      properties:
        id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        name: { type: string }
    TagInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
    ContactTag:
      type: object
      description: A tagging join row (contact x tag).
      properties:
        id: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        tag_id: { type: string, format: uuid }
        contact_vcard_uid: { type: string, format: uuid }
    ContactTagInput:
      type: object
      required: [contact_vcard_uid]
      properties:
        contact_vcard_uid: { type: string, format: uuid }

    # --- Relationship edges ---
...